+++ Shared/EventHandlers/Twitch.cs
@@ -26,6 +26,8 @@ namespace Slipstream.Shared.EventHandlers
 
         public delegate void OnTwitchReceivedWhisperHandler(EventHandler source, EventHandlerArgs<TwitchReceivedWhisper> e);
 
+        public delegate void OnTwitchUserSubscribedHandler(EventHandler source, EventHandlerArgs<TwitchUserSubscribed> e);
+
         public event OnTwitchCommandSendMessageHandler? OnTwitchCommandSendMessage;
 
         public event OnTwitchCommandSendWhisperHandler? OnTwitchCommandSendWhisper;
@@ -37,6 +39,8 @@ namespace Slipstream.Shared.EventHandlers
         public event OnTwitchReceivedMessageHandler? OnTwitchReceivedMessage;
 
         public event OnTwitchReceivedWhisperHandler? OnTwitchReceivedWhisper;
+
+        public event OnTwitchUserSubscribedHandler? OnTwitchUserSubscribed;
 
         public IEventHandler.HandledStatus HandleEvent(IEvent @event)
         {
@@ -96,6 +100,16 @@ namespace Slipstream.Shared.EventHandlers
                     if (OnTwitchReceivedWhisper != null)
                     {
                         OnTwitchReceivedWhisper.Invoke(Parent, new EventHandlerArgs<TwitchReceivedWhisper>(tev));
+                        return IEventHandler.HandledStatus.Handled;
+                    }
+                    else
+                    {
+                        return IEventHandler.HandledStatus.UseDefault;
+                    }
+                case TwitchUserSubscribed tev:
+                    if (OnTwitchUserSubscribed != null)
+                    {
+                        OnTwitchUserSubscribed.Invoke(Parent, new EventHandlerArgs<TwitchUserSubscribed>(tev));
                         return IEventHandler.HandledStatus.Handled;
                     }
                     else