+++ Shared/Factories/ITwitchEventFactory.cs
@@ -7,10 +7,17 @@ namespace Slipstream.Shared.Factories
     public interface ITwitchEventFactory
     {
         TwitchCommandSendMessage CreateTwitchCommandSendMessage(string message);
+
         TwitchConnected CreateTwitchConnected();
+
         TwitchDisconnected CreateTwitchDisconnected();
+
         TwitchReceivedMessage CreateTwitchReceivedMessage(string from, string message, bool moderator, bool subscriber, bool vip, bool broadcaster);
+
         TwitchReceivedWhisper CreateTwitchReceivedWhisper(string from, string message);
+
         TwitchCommandSendWhisper CreateTwitchCommandSendWhisper(string to, string message);
+
+        TwitchUserSubscribed CreateTwitchUserSubscribed(string name, string message, string subscriptionPlan, long months, string systemMessage);
     }
 }