+++ CHANGELOG.md
@@ -20,6 +20,7 @@
     - Removed events: IRacingCommandSendCurrentSession, IRacingCurrentSession and IRacingSessionState
     - Replaced with: IRacingPractice, IRacingQualify, IRacingRace, IRacingTesting, IRacingWarmup
     - If your script needs to get the current state, it can still use the lua function: iracing:send_session_state()
+  - TwitchPlugin: Adds events TwitchUserSubscribed (both sub and resub)
 
 ## [0.4.0](https://github.com/dennis/slipstream/releases/tag/v0.3.0) (2020-01-10)
 [Full Changelog](https://github.com/dennis/slipstream/compare/v0.3.0...v0.4.0)